feat: add nodeaffinity initialization from a kubeconfig file
This adds an optional input argument ``--kubeconfig` to the nodeaffinity main, which accepts a path to a local kube config file (for example, ~/.kube/config).  If specified, this makes the nodeaffinity injector use the kubeconfig file to connect to the kubernetes cluster.  The purpose of this is for local testing.